+++ flang/docs/ReleaseNotes.md
@@ -36,6 +36,14 @@ page](https://llvm.org/releases/).
 
 ## Build System Changes
 
+ * The Fortran Runtime library has been move to a new top-level directory
+   named "flang-rt". The library was also renamed from `libFortranRuntime.a` to
+   `libflang_rt.a`. It now supports the
+   LLVM_ENABLE_RUNTIMES mechanism to build flang-rt for multiple target
+   triples. libflang_rt.a will now be emitted into Clang's per-target
+   resource directory (next to libclang_rt.*.*) where it is also found by
+   Flang's driver.
+
 ## New Issues Found